backend/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9

# Pydantic & settings
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AWS S3
boto3==1.34.144

# Image processing
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
